move ExecutionWitness to stateless.py
